+++ NEWS.md
@@ -1,4 +1,8 @@
 # strayr 0.0.0.901
+* strayr() now recognises 'Australia', 'Other', and 'Total' as distinct strings for matching
+* Extraneous elements (such as "the") no longer removed before fuzzy matching
+
+# strayr 0.0.0.901
 
 * Conversion to and from `code` is now possible; see [the ABS website](https://www.abs.gov.au/ausstats/abs@.nsf/Lookup/by%20Subject/1270.0.55.001~July%202016~Main%20Features~Australia%20(AUS)%20and%20State%20%7C%20Territory%20(S%7CT)~10017) for codes